DjangoIntegration1st/requirements.txt
Django==5.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
openai==1.23.2
httpx==0.27.0
python-dotenv==1.0.1
firebase-admin==6.4.0
psycopg[binary]==3.1.18
redis==5.0.1
django-redis==5.4.0
channels==4.1.0
daphne==4.0.0
twisted>=22.4
Pillow==10.2.0
